generation-startups (server)/blogapi-master/blog/requirements.txt
Django==3.1.7
transformers~=4.25.1
djangorestframework~=3.12.4
torch~=1.13.0